Move db.js
