+++ src/modules/hearts.js
@@ -1,4 +1,5 @@
-const { db } = require('./db');
+const { db } = require('../db');
+
 const { getMonthStart, getPrevMonthEnd } = require('../utils');
 
 const {
