+++ src/modules/things.js
@@ -1,4 +1,5 @@
-const { db } = require('./db');
+const { db } = require('../db');
+
 const { thingsPollLength, thingsMinVotesScalar } = require('../config');
 
 const Polls = require('./polls');
