+++ test/admin.test.js
@@ -6,7 +6,7 @@ chai.use(chaiAsPromised);
 
 const { HOUR, DAY } = require('../src/constants');
 const { getMonthStart, getMonthEnd, getNextMonthStart, getPrevMonthEnd, getDateStart } = require('../src/utils');
-const { db } = require('../src/modules/db');
+const { db } = require('../src/db');
 
 const Admin = require('../src/modules/admin');
